backend/requirements.txt
flask
flask-cors
openai-whisper
google-generativeai
pydub
python-dotenv
torch
torchaudio
noisereduce
soundfile
gunicorn
